+++ Samples/iOS-Swift/iOS-Swift/ViewControllers/TraceTestViewController.swift
@@ -72,7 +72,7 @@ class TraceTestViewController: UIViewController {
                 return
             }
 
-            UIAssert.isEqual(child.data?["url"] as? String, "/sentry-logo-black.png", "Could not read url data value")
+            UIAssert.isEqual(child.data["url"] as? String, "/sentry-logo-black.png", "Could not read url data value")
 
             UIAssert.isEqual(child.tags["http.status_code"], "200", "Could not read status_code tag value")