+++ Sources/Sentry/SentrySpanContext.m
@@ -1,5 +1,6 @@
 #import "SentrySpanContext.h"
 #import "SentryId.h"
+#import "SentryLog.h"
 #import "SentrySpanId.h"
 
 NS_ASSUME_NONNULL_BEGIN
@@ -48,6 +49,11 @@ NS_ASSUME_NONNULL_BEGIN
         _sampled = sampled;
         _operation = operation;
         _spanDescription = description;
+
+        SENTRY_LOG_DEBUG(
+            @"Created span context with trace ID %@; span ID %@; parent span ID %@; operation %@",
+            traceId.sentryIdString, spanId.sentrySpanIdString, parentId.sentrySpanIdString,
+            operation);
     }
     return self;
 }