+++ Tests/SentryTests/Integrations/Performance/Network/SentryNetworkTrackerTests.swift
@@ -456,9 +456,6 @@ class SentryNetworkTrackerTests: XCTestCase {
     
     // Although we only run this test above the below specified versions, we expect the
     // implementation to be thread safe
-    @available(tvOS 10.0, *)
-    @available(OSX 10.12, *)
-    @available(iOS 10.0, *)
     func testResumeCalledMultipleTimesConcurrent_OneSpanCreated() {
         let task = createDataTask()
         let sut = fixture.getSut()
@@ -484,9 +481,6 @@ class SentryNetworkTrackerTests: XCTestCase {
     
     // Although we only run this test above the below specified versions, we expect the
     // implementation to be thread safe
-    @available(tvOS 10.0, *)
-    @available(OSX 10.12, *)
-    @available(iOS 10.0, *)
     func testChangeStateMultipleTimesConcurrent_OneSpanFinished() {
         let task = createDataTask()
         let sut = fixture.getSut()
@@ -752,9 +746,9 @@ class SentryNetworkTrackerTests: XCTestCase {
             XCTAssertNil(httpStatusCode)
         }
         
-        let path = span.data!["url"] as? String
-        let method = span.data!["method"] as? String
-        let requestType = span.data!["type"] as? String
+        let path = span.data["url"] as? String
+        let method = span.data["method"] as? String
+        let requestType = span.data["type"] as? String
         
         XCTAssertEqual(path, task.currentRequest!.url!.path)
         XCTAssertEqual(method, task.currentRequest!.httpMethod)