+++ CHANGELOG.md
@@ -13,10 +13,12 @@ This version adds a dependency on Swift.
 - Remove `- [SentryOptions initWithDict:didFailWithError:]` (#2404)
 - Rename `- [SentrySDK startWithOptionsObject:]` to `- [SentrySDK startWithOptions:]` (#2404)
 - Remove `- [SentryOptions sdkInfo]` (#2404)
+- Make `SpanProtocol.data` non nullable (#2409)
 - Mark `- [SpanProtocol setExtraValue:forKey:]` as deprecated (#2413)
 - Make SpanContext immutable (#2408)
     - Remove tags from SpanContext 
     - Remove context property from SentrySpan
+- Bump minimum supported OS versions to macOS 10.13, iOS 11, tvOS 11, and watchOS 4 (#2414)
 
 ## 7.31.2